SSH debugging

+++ scripts/02-create-source-images/02-prepare-ssh.sh
@@ -8,39 +8,34 @@
 # If this is a hosted build agent, this SSH public/private key pair's lifetime is only as long as the build agent's.
 
 # Clean up existing key files, if any
-delCmd="rm ./""$DEPLOYMENT_SSH_USER_KEY_NAME""*"
-#echo $delCmd
-eval $delCmd
-delCmd="rm ~/.ssh/""$DEPLOYMENT_SSH_USER_KEY_NAME""*"
-#echo $delCmd
-eval $delCmd
+if [[ -f ~/.ssh/"$DEPLOYMENT_SSH_USER_KEY_NAME" ]]
+then
+	delCmd="rm ~/.ssh/""$DEPLOYMENT_SSH_USER_KEY_NAME""*"
+	#echo $delCmd
+	eval $delCmd
+fi
+
 
 # Generate new deployment user public and private key pair and write the files here
-ssh-keygen -q -m "PEM" -f "./""$DEPLOYMENT_SSH_USER_KEY_NAME" -t "$DEPLOYMENT_SSH_KEY_TYPE" -b $DEPLOYMENT_SSH_KEY_BITS -N "$DEPLOYMENT_SSH_KEY_PASSPHRASE" -C "$DEPLOYMENT_SSH_USER_NAME"
+ssh-keygen -q -m "PEM" -f ~/.ssh/"$DEPLOYMENT_SSH_USER_KEY_NAME" -t "$DEPLOYMENT_SSH_KEY_TYPE" -b $DEPLOYMENT_SSH_KEY_BITS -N "$DEPLOYMENT_SSH_KEY_PASSPHRASE" -C "$DEPLOYMENT_SSH_USER_NAME"
 
 # Also write a file for the admin public key
-echo $VM_ADMIN_SSH_PUBLIC_KEY > "./""$VM_ADMIN_SSH_USER_KEY_NAME"".pub"
+echo $VM_ADMIN_SSH_PUBLIC_KEY > ~/.ssh/"$VM_ADMIN_SSH_USER_KEY_NAME".pub
 
-# Move SSH key files to ~/.ssh
-mkdir -p ~/.ssh
-mv "./""$DEPLOYMENT_SSH_USER_KEY_NAME" ~/.ssh
-cp "./""$DEPLOYMENT_SSH_USER_KEY_NAME"".pub" ~/.ssh
-cp "./""$VM_ADMIN_SSH_USER_KEY_NAME"".pub" ~/.ssh
+## Set SSH key file permissions
+## Private key - restrictive
+#privCmd="chmod 600 ~/.ssh/""$DEPLOYMENT_SSH_USER_KEY_NAME"
+##echo $privCmd
+#eval $privCmd
 
-# Set SSH key file permissions
-# Private key - restrictive
-privCmd="chmod 600 ~/.ssh/""$DEPLOYMENT_SSH_USER_KEY_NAME"
-#echo $privCmd
-eval $privCmd
+## Public keys - less restrictive
+#pubCmd="chmod 644 ~/.ssh/""$DEPLOYMENT_SSH_USER_KEY_NAME"".pub"
+##echo $pubCmd
+#eval $pubCmd
 
-# Public keys - less restrictive
-pubCmd="chmod 644 ~/.ssh/""$DEPLOYMENT_SSH_USER_KEY_NAME"".pub"
-#echo $pubCmd
-eval $pubCmd
-
-pubCmd="chmod 644 ~/.ssh/""$VM_ADMIN_SSH_USER_KEY_NAME"".pub"
-#echo $pubCmd
-eval $pubCmd
+#pubCmd="chmod 644 ~/.ssh/""$VM_ADMIN_SSH_USER_KEY_NAME"".pub"
+##echo $pubCmd
+#eval $pubCmd
 
 # Add deployment private SSH key to SSH agent
 eval $(ssh-agent)
